Lotto_ml_web/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
requests>=2.31.0
joblib>=1.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
